refactor(types): extract IPC channel enums into dependency-free channels.ts
Move the Main/ToMain enums (and the MAIN const) into src/types/IPC/channels.ts so consumers
can import channel identifiers without dragging in the whole payload/Electron type graph;
Main.ts/ToMain.ts re-export them. Also decouple the type layer from Electron and the
frontend: inline AutosizeTypes in Show.ts, and add local Rectangle + DataFolderName in
Main.ts. Adds the YJS transport channel.
