+++ src/types/IPC/Main.ts
@@ -18,162 +18,10 @@ import type { Event } from "./../Calendar"
 import type { History } from "./../History"
 import type { SaveData, SaveListSyncedSettings } from "./../Save"
 
-export const MAIN = "MAIN"
-
-export enum Main {
-    // DEV
-    LOG = "LOG",
-    IS_DEV = "IS_DEV",
-    GET_CACHE_PATH = "GET_CACHE_PATH",
-    // APP
-    VERSION = "VERSION",
-    GET_OS = "GET_OS",
-    DEVICE_ID = "DEVICE_ID",
-    GET_DEVICE_NAME = "GET_DEVICE_NAME",
-    IP = "IP",
-    CHECK_RAM_USAGE = "CHECK_RAM_USAGE",
-    // STORES
-    SETTINGS = "SETTINGS",
-    SYNCED_SETTINGS = "SYNCED_SETTINGS",
-    STAGE = "STAGE",
-    PROJECTS = "PROJECTS",
-    OVERLAYS = "OVERLAYS",
-    TEMPLATES = "TEMPLATES",
-    EVENTS = "EVENTS",
-    MEDIA = "MEDIA",
-    THEMES = "THEMES",
-    DRIVE_API_KEY = "DRIVE_API_KEY",
-    HISTORY = "HISTORY",
-    USAGE = "USAGE",
-    CACHE = "CACHE",
-    // WINDOW
-    CLOSE = "CLOSE",
-    MAXIMIZE = "MAXIMIZE",
-    MAXIMIZED = "MAXIMIZED",
-    MINIMIZE = "MINIMIZE",
-    FULLSCREEN = "FULLSCREEN",
-    /////
-    IMPORT = "IMPORT",
-    IMPORT_FILES = "IMPORT_FILES",
-    BIBLE = "BIBLE",
-    SHOW = "SHOW",
-    SAVE = "SAVE",
-    BACKUPS = "BACKUPS",
-    DELETE_BACKUP = "DELETE_BACKUP",
-    ///////////////////
-    SPELLCHECK = "SPELLCHECK",
-    ////
-    SHOWS = "SHOWS",
-    AUTO_UPDATE = "AUTO_UPDATE",
-    URL = "URL",
-    LANGUAGE = "LANGUAGE",
-    GET_PATHS = "GET_PATHS",
-    DATA_PATH = "DATA_PATH",
-    UPDATE_DATA_PATH = "UPDATE_DATA_PATH",
-    LOG_ERROR = "LOG_ERROR",
-    OPEN_LOG = "OPEN_LOG",
-    OPEN_CACHE = "OPEN_CACHE",
-    OPEN_APPDATA = "OPEN_APPDATA",
-    OPEN_FOLDER_PATH = "OPEN_FOLDER_PATH",
-    OPEN_NOW_PLAYING = "OPEN_NOW_PLAYING",
-    GET_STORE_VALUE = "GET_STORE_VALUE",
-    SET_STORE_VALUE = "SET_STORE_VALUE",
-    DELETE_SHOWS = "DELETE_SHOWS",
-    DELETE_SHOWS_NI = "DELETE_SHOWS_NI",
-    REFRESH_SHOWS = "REFRESH_SHOWS",
-    GET_EMPTY_SHOWS = "GET_EMPTY_SHOWS",
-    FULL_SHOWS_LIST = "FULL_SHOWS_LIST",
-    GET_SCREENS = "GET_SCREENS",
-    GET_WINDOWS = "GET_WINDOWS",
-    GET_DISPLAYS = "GET_DISPLAYS",
-    OUTPUT = "OUTPUT",
-    DOES_MEDIA_EXIST = "DOES_MEDIA_EXIST",
-    GET_THUMBNAIL = "GET_THUMBNAIL",
-    SAVE_IMAGE = "SAVE_IMAGE",
-    PDF_TO_IMAGE = "PDF_TO_IMAGE",
-    READ_EXIF = "READ_EXIF",
-    MEDIA_CODEC = "MEDIA_CODEC",
-    MEDIA_TRACKS = "MEDIA_TRACKS",
-    DOWNLOAD_LESSONS_MEDIA = "DOWNLOAD_LESSONS_MEDIA",
-    MEDIA_DOWNLOAD = "MEDIA_DOWNLOAD",
-    MEDIA_IS_DOWNLOADED = "MEDIA_IS_DOWNLOADED",
-    NOW_PLAYING = "NOW_PLAYING",
-    NOW_PLAYING_UNSET = "NOW_PLAYING_UNSET",
-    // MEDIA_BASE64 = "MEDIA_BASE64",
-    READ_AUDIO_METADATA = "READ_AUDIO_METADATA",
-    CAPTURE_SLIDE = "CAPTURE_SLIDE",
-    ACCESS_CAMERA_PERMISSION = "ACCESS_CAMERA_PERMISSION",
-    ACCESS_MICROPHONE_PERMISSION = "ACCESS_MICROPHONE_PERMISSION",
-    ACCESS_SCREEN_PERMISSION = "ACCESS_SCREEN_PERMISSION",
-    LIBREOFFICE_CONVERT = "LIBREOFFICE_CONVERT",
-    SLIDESHOW_GET_APPS = "SLIDESHOW_GET_APPS",
-    START_SLIDESHOW = "START_SLIDESHOW",
-    PRESENTATION_CONTROL = "PRESENTATION_CONTROL",
-    START = "START",
-    STOP = "STOP",
-    SERVER_DATA = "SERVER_DATA",
-    WEBSOCKET_START = "WEBSOCKET_START",
-    WEBSOCKET_STOP = "WEBSOCKET_STOP",
-    API_TRIGGER = "API_TRIGGER",
-    EMIT_OSC = "EMIT_OSC",
-    GET_MIDI_OUTPUTS = "GET_MIDI_OUTPUTS",
-    GET_MIDI_INPUTS = "GET_MIDI_INPUTS",
-    SEND_MIDI = "SEND_MIDI",
-    RECEIVE_MIDI = "RECEIVE_MIDI",
-    CLOSE_MIDI = "CLOSE_MIDI",
-    GET_LYRICS = "GET_LYRICS",
-    SEARCH_LYRICS = "SEARCH_LYRICS",
-    RECORDER = "RECORDER",
-    RESTORE = "RESTORE",
-    SYSTEM_OPEN = "SYSTEM_OPEN",
-    LOCATE_MEDIA_FILE = "LOCATE_MEDIA_FILE",
-    GET_MEDIA_FOLDER_PATH = "GET_MEDIA_FOLDER_PATH",
-    SET_MEDIA_FOLDER_PATH = "SET_MEDIA_FOLDER_PATH",
-    GET_SIMILAR = "GET_SIMILAR",
-    BUNDLE_MEDIA_FILES = "BUNDLE_MEDIA_FILES",
-    MEDIA_FOLDER_COPY = "MEDIA_FOLDER_COPY",
-    READ_BIBLES_FOLDER = "READ_BIBLES_FOLDER",
-    FILE_INFO = "FILE_INFO",
-    READ_FOLDER = "READ_FOLDER",
-    READ_FILE = "READ_FILE",
-    OPEN_FOLDER = "OPEN_FOLDER",
-    OPEN_FILE = "OPEN_FILE",
-    // SYNC
-    CAN_SYNC = "CAN_SYNC",
-    GET_TEAMS = "GET_TEAMS",
-    CLOUD_DATA = "CLOUD_DATA",
-    CLOUD_CHANGED = "CLOUD_CHANGED",
-    CLOUD_SYNC = "CLOUD_SYNC",
-    RESTORE_CLOUD_BACKUP = "RESTORE_CLOUD_BACKUP",
-    GET_CONVERSATION_ID = "GET_CONVERSATION_ID",
-    SEND_SOCKET_MESSAGE = "SEND_SOCKET_MESSAGE",
-    // Provider-based routing
-    PROVIDER_LOAD_SERVICES = "PROVIDER_LOAD_SERVICES",
-    PROVIDER_DISCONNECT = "PROVIDER_DISCONNECT",
-    PROVIDER_STARTUP_LOAD = "PROVIDER_STARTUP_LOAD",
-    PROVIDER_FETCH_FOLDERS = "PROVIDER_FETCH_FOLDERS",
-    PCO_LIVE_GET = "PCO_LIVE_GET",
-    PCO_PUSHER_AUTH = "PCO_PUSHER_AUTH",
-    PCO_FETCH_SERVICE_TREE = "PCO_FETCH_SERVICE_TREE",
-    PCO_LOAD_PLAN = "PCO_LOAD_PLAN",
-    // Content Library
-    GET_CONTENT_PROVIDERS = "GET_CONTENT_PROVIDERS",
-    GET_CONTENT_LIBRARY = "GET_CONTENT_LIBRARY",
-    GET_PROVIDER_CONTENT = "GET_PROVIDER_CONTENT",
-    CHECK_MEDIA_LICENSE = "CHECK_MEDIA_LICENSE",
-    // Timecode
-    TIMECODE_START = "TIMECODE_START",
-    TIMECODE_STOP = "TIMECODE_STOP",
-    TIMECODE_VALUE = "TIMECODE_VALUE",
-    TIMECODE_AUDIO_DATA = "TIMECODE_AUDIO_DATA",
-    TIMECODE_STATUS = "TIMECODE_STATUS",
-    // Spotify
-    SPOTIFY_GET_STATE = "SPOTIFY_GET_STATE",
-    SPOTIFY_COMMAND = "SPOTIFY_COMMAND",
-    // FFmpeg Download
-    FFMPEG_CHECK = "FFMPEG_CHECK",
-    FFMPEG_DOWNLOAD = "FFMPEG_DOWNLOAD"
-}
+// enum + MAIN const are declared in ./channels (no imports) and re-exported here
+// so the channel enum can be consumed without pulling in the payload/Electron types below.
+export { MAIN, Main } from "./channels"
+import { Main } from "./channels"
 
 export interface MainSendPayloads {
     // DEV
@@ -239,6 +87,7 @@ export interface MainSendPayloads {
     [Main.FILE_INFO]: string
     [Main.READ_FOLDER]: { path: string | string[]; depth?: number; generateThumbnails?: boolean; captureFolderContent?: boolean }
     [Main.READ_FILE]: { path: string }
+    [Main.CREATE_FOLDER]: { path: string; name: string }
     [Main.OPEN_FOLDER]: { channel: string; title?: string; path?: string }
     [Main.OPEN_FILE]: { id: string; channel: string; title?: string; filter: any; multiple: boolean; read?: boolean }
     // SYNC
@@ -344,6 +193,7 @@ export interface MainReturnPayloads {
     [Main.READ_BIBLES_FOLDER]: { path: string; name: string }[]
     [Main.FILE_INFO]: { path: string; stat: Stats; extension: string; folder: boolean } | null
     [Main.READ_FOLDER]: Promise<{ [key: string]: FileFolder }>
+    [Main.CREATE_FOLDER]: string
     [Main.READ_FILE]: { content: string }
     // SYNC
     [Main.CAN_SYNC]: Promise<boolean>
